+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Updated the `Nft` class to include the contract metadata in the `Nft.contract` field.
 - Added commonly used utility methods from ethers.js into a top-level `Utils` export.
 
+## Minor Changes
+
+- Added support for `withTokenBalances` in `NftNamespace.getOwnersForContract()`.
+
 ## 2.0.4
 
 ### Minor Changes